+++ src/components/Screen.jsx
@@ -3,7 +3,7 @@ import React from 'react'
 function Screen({ value }) {
     let screenContent = 'ERROR'
 
-    if (parseInt(value, 10) >= 0 && parseInt(value, 10) <= 999999999 ) {
+    if (parseFloat(value, 10) >= 0 && parseFloat(value, 10) <= 999999999 ) {
         screenContent = value
     }
